make sure glad is the only GL loader
