+++ src/imgui_runner/runner_glfw.cpp
@@ -2,13 +2,16 @@
 #include "runner_glfw.h"
 
 #include "imgui_runner/runner_glfw.h"
+#include "imgui_runner/glad_callbacks.h"
 #include <GLFW/glfw3.h>
 #include <examples/imgui_impl_glfw.h>
 #include <examples/imgui_impl_opengl3.h>
 #include <glad/glad.h>
 #include <imgui.h>
 #include <imgui_impl_glfw.h>
+
 #include <stdexcept>
+#include <sstream>
 
 // SDL + OpenGL3 + glad
 
@@ -91,27 +94,31 @@ void RunnerGlfw::CreateWindowAndContext()
 void RunnerGlfw::InitGlLoader()
 {
 #ifndef __EMSCRIPTEN__
-  if (!gladLoadGLLoader((GLADloadproc)glfwGetProcAddress))
-    throw std::runtime_error("gladLoadGLLoader: Failed");
-  if (!GLAD_GL_VERSION_3_3)
-    throw(std::runtime_error("GLAD could not initialize OpenGl 3.3"));
+//  if (!gladLoadGLLoader((GLADloadproc)glfwGetProcAddress))
+//    throw std::runtime_error("gladLoadGLLoader: Failed");
+//  if (!GLAD_GL_VERSION_3_3)
+//    throw(std::runtime_error("GLAD could not initialize OpenGl 3.3"));
 
-//  // Initialize OpenGL loader
-//#if defined(IMGUI_IMPL_OPENGL_LOADER_GL3W)
-//  bool err = gl3wInit() != 0;
-//#elif defined(IMGUI_IMPL_OPENGL_LOADER_GLEW)
-//  bool err = glewInit() != GLEW_OK;
-//#elif defined(IMGUI_IMPL_OPENGL_LOADER_GLAD)
-//  bool err = gladLoadGL() == 0;
-//#else
-//  bool err = false; // If you use IMGUI_IMPL_OPENGL_LOADER_CUSTOM, your loader is likely to requires some form of initialization.
-//#endif
-//  if (err)
-//  {
-//    fprintf(stderr, "Failed to initialize OpenGL loader!\n");
-//    return 1;
-//  }
+  // Initialize OpenGL loader
+#if defined(IMGUI_IMPL_OPENGL_LOADER_GL3W)
+  bool err = gl3wInit() != 0;
+#elif defined(IMGUI_IMPL_OPENGL_LOADER_GLEW)
+  bool err = glewInit() != GLEW_OK;
+#elif defined(IMGUI_IMPL_OPENGL_LOADER_GLAD)
+  bool err = gladLoadGLLoader((GLADloadproc)glfwGetProcAddress) == 0;
+#else
+  bool err = false; // If you use IMGUI_IMPL_OPENGL_LOADER_CUSTOM, your loader is likely to requires some form of initialization.
+#endif
+  if (err)
+  {
+    throw std::runtime_error("RunnerGlfw::InitGlLoader(): Failed to initialize OpenGL loader!");
+  }
 #endif // #ifndef __EMSCRIPTEN__
+
+#ifdef GLAD_DEBUG
+  glad_set_pre_callback(glad_pre_call_callback);
+  glad_set_post_callback(glad_post_call_callback);
+#endif
 }
 
 void RunnerGlfw::SetupPlatformRendererBindings()
